организация кода
